Refactored example code and turned the message function into a class.  Then made it so on click the timestamp is messaged to the console.
