+++ server/public/styles.css
@@ -25,5 +25,4 @@ ul {
 
 .timestamp {
   cursor: pointer;
-  color: blue;
 }
